refactor(assets): restructure design-system folder
