+++ src/assets/_design-system.scss
@@ -1,2 +1,2 @@
-@import 'designSystem/variables';
-@import 'designSystem/mixins';
+@import 'design-system/variables/variables';
+@import 'design-system/mixins/mixins';
