+++ src/assets/design-system/docs/components/ColorPalette.vue
