+++ src/assets/design-system/variables/_color-palette.scss
