+++ src/assets/design-system/variables/components/atoms/_checkbox.scss
