+++ src/assets/design-system/variables/components/atoms/_input.scss
