+++ src/assets/design-system/variables/components/atoms/_loader.scss
