+++ src/assets/design-system/variables/components/atoms/_select.scss
