+++ src/assets/design-system/variables/components/atoms/_slider.scss
