+++ src/assets/design-system/variables/components/atoms/_star-rating.scss
