+++ src/assets/design-system/variables/components/atoms/_toggle.scss
