+++ src/assets/design-system/variables/components/data-display/_badge.scss
