+++ src/assets/design-system/variables/components/molecules/_back-to-top.scss
