+++ src/assets/design-system/variables/components/molecules/_card.scss
