+++ src/assets/design-system/variables/components/molecules/_carousel.scss
