+++ src/assets/design-system/variables/components/molecules/_modal.scss
