+++ src/assets/design-system/variables/components/molecules/_notification-stack.scss
