+++ src/assets/design-system/variables/components/molecules/_pagination.scss
