+++ src/assets/design-system/variables/components/molecules/_tooltip.scss
