+++ src/assets/design-system/variables/components/organisms/_accordion.scss
