+++ src/assets/design-system/variables/components/organisms/_autocomplete.scss
