+++ src/assets/design-system/variables/components/organisms/_calendar.scss
