+++ src/assets/design-system/variables/components/organisms/_data-table.scss
