+++ src/assets/design-system/variables/components/organisms/_footer.scss
