+++ src/assets/design-system/variables/components/organisms/_nav-bar.scss
