+++ src/assets/design-system/variables/components/organisms/_sidebar.scss
