+++ src/assets/design-system/variables/components/organisms/_tab-group.scss
