+++ src/assets/reset.scss
@@ -1,5 +1,5 @@
-@import 'designSystem/variables';
-@import 'designSystem/mixins';
+@import 'design-system/variables/variables';
+@import 'design-system/mixins/mixins';
 
 /*! minireset.css v0.0.6 | MIT License | github.com/jgthms/minireset.css */
 html,
