+++ src/assets/typography.scss
@@ -1,5 +1,5 @@
-@import 'designSystem/variables';
-@import 'designSystem/mixins';
+@import 'design-system/variables/variables';
+@import 'design-system/mixins/mixins';
 
 html {
   font-size: 62.5%;
