+++ src/components/atoms/icons/Icons.stories.ts
@@ -1,5 +1,5 @@
 import { storiesOf } from '@storybook/vue';
-import IconList from '@/assets/designSystem/docs/components/IconList.vue';
+import IconList from '@/assets/design-system/docs/components/IconList.vue';
 
 const story = storiesOf('Foundation|General', module) as any;
